[e2e] Move migration tests to their own package
