+++ tests/hotplug/cpu.go
@@ -6,6 +6,8 @@ import (
 	"fmt"
 	"time"
 
+	"kubevirt.io/kubevirt/tests/libmigration"
+
 	"github.com/onsi/gomega/gstruct"
 	gomegatypes "github.com/onsi/gomega/types"
 
@@ -148,7 +150,7 @@ var _ = Describe("[sig-compute][Serial]CPU Hotplug", decorators.SigCompute, deco
 				}
 				return false
 			}, 30*time.Second, time.Second).Should(BeTrue())
-			tests.ExpectMigrationSuccess(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.ExpectMigrationSuccess(virtClient, migration, tests.MigrationWaitTime)
 
 			By("Ensuring the libvirt domain has 4 enabled cores")
 			Eventually(func() cpuCount {
@@ -237,7 +239,7 @@ var _ = Describe("[sig-compute][Serial]CPU Hotplug", decorators.SigCompute, deco
 			_, err = virtClient.VirtualMachine(vm.Namespace).Patch(context.Background(), vm.Name, types.JSONPatchType, patchData, &k8smetav1.PatchOptions{})
 			Expect(err).To(HaveOccurred())
 
-			tests.ExpectMigrationSuccess(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.ExpectMigrationSuccess(virtClient, migration, tests.MigrationWaitTime)
 
 			By("Enabling the second socket")
 			patchData, err = patch.GenerateTestReplacePatch("/spec/template/spec/domain/cpu/sockets", 1, 2)
@@ -265,7 +267,7 @@ var _ = Describe("[sig-compute][Serial]CPU Hotplug", decorators.SigCompute, deco
 				}
 				return false
 			}, 30*time.Second, time.Second).Should(BeTrue())
-			tests.ExpectMigrationSuccess(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.ExpectMigrationSuccess(virtClient, migration, tests.MigrationWaitTime)
 
 			By("Ensuring the libvirt domain has 4 enabled cores")
 			Eventually(func() cpuCount {
