+++ tests/hyperv_test.go
@@ -6,6 +6,10 @@ import (
 	"strings"
 	"time"
 
+	"kubevirt.io/kubevirt/tests/libmigration"
+
+	"kubevirt.io/kubevirt/tests/libinfra"
+
 	"kubevirt.io/kubevirt/tests/decorators"
 	"kubevirt.io/kubevirt/tests/framework/kubevirt"
 
@@ -85,10 +89,10 @@ var _ = Describe("[Serial][sig-compute] Hyper-V enlightenments", Serial, decorat
 
 				By("Migrating the VM")
 				migration := tests.NewRandomMigration(reEnlightenmentVMI.Name, reEnlightenmentVMI.Namespace)
-				migrationUID := tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migrationUID := libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				By("Checking VMI, confirm migration state")
-				tests.ConfirmVMIPostMigration(virtClient, reEnlightenmentVMI, migrationUID)
+				libmigration.ConfirmVMIPostMigration(virtClient, reEnlightenmentVMI, migrationUID)
 			})
 
 			It("should have TSC frequency set up in label and domain", func() {
@@ -128,7 +132,7 @@ var _ = Describe("[Serial][sig-compute] Hyper-V enlightenments", Serial, decorat
 			BeforeEach(func() {
 				if isTSCFrequencyExposed(virtClient) {
 					for _, node := range libnode.GetAllSchedulableNodes(virtClient).Items {
-						stopNodeLabeller(node.Name, virtClient)
+						libinfra.StopNodeLabeller(node.Name, virtClient)
 						removeTSCFrequencyFromNode(node)
 					}
 				}
@@ -136,7 +140,7 @@ var _ = Describe("[Serial][sig-compute] Hyper-V enlightenments", Serial, decorat
 
 			AfterEach(func() {
 				for _, node := range libnode.GetAllSchedulableNodes(virtClient).Items {
-					_ = resumeNodeLabeller(node.Name, virtClient)
+					_ = libinfra.ResumeNodeLabeller(node.Name, virtClient)
 				}
 			})
