+++ tests/kubectl_test.go
@@ -5,6 +5,8 @@ import (
 	"strings"
 	"time"
 
+	"kubevirt.io/kubevirt/tests/libmigration"
+
 	"kubevirt.io/kubevirt/tests/decorators"
 
 	"kubevirt.io/kubevirt/tests/framework/checks"
@@ -190,7 +192,7 @@ var _ = Describe("[sig-compute]oc/kubectl integration", decorators.SigCompute, f
 				}, tests.MigrationWaitTime, 1*time.Second).Should(Succeed(), "migration creation should succeed")
 				migration = migrationCreated
 
-				tests.ExpectMigrationSuccess(virtClient, migration, tests.MigrationWaitTime)
+				libmigration.ExpectMigrationSuccess(virtClient, migration, tests.MigrationWaitTime)
 
 				k8sClient := clientcmd.GetK8sCmdClient()
 				result, _, err := clientcmd.RunCommand(k8sClient, "get", "vmim", migration.Name)
