+++ tests/migration/BUILD.bazel
@@ -3,40 +3,29 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "datavolume.go",
-        "events.go",
-        "export.go",
         "framework.go",
-        "guestfs.go",
-        "hotplug.go",
-        "imageupload.go",
-        "memorydump.go",
-        "reservation.go",
-        "restore.go",
-        "snapshot.go",
-        "storage.go",
+        "migration.go",
+        "migration_policy.go",
     ],
-    importpath = "kubevirt.io/kubevirt/tests/storage",
+    importpath = "kubevirt.io/kubevirt/tests/migration",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apimachinery/patch:go_default_library",
+        "//pkg/certificates/triple:go_default_library",
         "//pkg/certificates/triple/cert:go_default_library",
-        "//pkg/host-disk:go_default_library",
         "//pkg/pointer:go_default_library",
-        "//pkg/storage/reservation:go_default_library",
-        "//pkg/storage/types:go_default_library",
+        "//pkg/util/cluster:go_default_library",
+        "//pkg/util/hardware:go_default_library",
         "//pkg/virt-config:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter:go_default_library",
-        "//pkg/virt-operator/resource/generate/components:go_default_library",
-        "//pkg/virtctl/guestfs:go_default_library",
-        "//pkg/virtctl/vm:go_default_library",
-        "//staging/src/kubevirt.io/api/clone/v1alpha1:go_default_library",
-        "//staging/src/kubevirt.io/api/core:go_default_library",
+        "//pkg/virt-controller/services:go_default_library",
+        "//pkg/virt-controller/watch:go_default_library",
+        "//pkg/virt-controller/watch/topology:go_default_library",
+        "//pkg/virt-handler:go_default_library",
+        "//pkg/virt-handler/cgroup:go_default_library",
+        "//pkg/virt-handler/cmd-client:go_default_library",
+        "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
-        "//staging/src/kubevirt.io/api/export/v1alpha1:go_default_library",
-        "//staging/src/kubevirt.io/api/instancetype:go_default_library",
-        "//staging/src/kubevirt.io/api/instancetype/v1beta1:go_default_library",
-        "//staging/src/kubevirt.io/api/snapshot/v1alpha1:go_default_library",
+        "//staging/src/kubevirt.io/api/migrations/v1alpha1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//tests:go_default_library",
@@ -44,18 +33,19 @@ go_library(
         "//tests/console:go_default_library",
         "//tests/containerdisk:go_default_library",
         "//tests/decorators:go_default_library",
-        "//tests/errorhandling:go_default_library",
         "//tests/events:go_default_library",
         "//tests/exec:go_default_library",
         "//tests/flags:go_default_library",
         "//tests/framework/checks:go_default_library",
+        "//tests/framework/cleanup:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
-        "//tests/framework/storage:go_default_library",
         "//tests/libdv:go_default_library",
-        "//tests/libinstancetype:go_default_library",
+        "//tests/libinfra:go_default_library",
         "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
+        "//tests/libnet/job:go_default_library",
+        "//tests/libnet/service:go_default_library",
         "//tests/libnode:go_default_library",
         "//tests/libstorage:go_default_library",
         "//tests/libvmi:go_default_library",
@@ -63,32 +53,22 @@ go_library(
         "//tests/testsuite:go_default_library",
         "//tests/util:go_default_library",
         "//tests/watcher:go_default_library",
+        "//tools/vms-generator/utils:go_default_library",
         "//vendor/github.com/google/goexpect:go_default_library",
-        "//vendor/github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/github.com/onsi/gomega/format:go_default_library",
-        "//vendor/github.com/onsi/gomega/gstruct:go_default_library",
-        "//vendor/github.com/onsi/gomega/types:go_default_library",
-        "//vendor/github.com/openshift/api/route/v1:go_default_library",
-        "//vendor/github.com/pborman/uuid:go_default_library",
-        "//vendor/github.com/spf13/cobra:go_default_library",
-        "//vendor/k8s.io/api/admissionregistration/v1:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/api/networking/v1:go_default_library",
-        "//vendor/k8s.io/api/rbac/v1:go_default_library",
-        "//vendor/k8s.io/api/storage/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
+        "//vendor/k8s.io/api/policy/v1:go_default_library",
+        "//vendor/k8s.io/api/policy/v1beta1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/fields:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/labels:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/intstr:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/rand:go_default_library",
-        "//vendor/k8s.io/client-go/tools/remotecommand:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/strategicpatch:go_default_library",
         "//vendor/k8s.io/utils/pointer:go_default_library",
         "//vendor/kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1:go_default_library",
-        "//vendor/sigs.k8s.io/yaml:go_default_library",
     ],
 )
