+++ tests/migration/migration.go
@@ -17,7 +17,7 @@
  *
  */
 
-package tests_test
+package migration
 
 import (
 	"context"
@@ -32,6 +32,8 @@ import (
 	"strings"
 	"sync"
 	"time"
+
+	"kubevirt.io/kubevirt/tests/libmigration"
 
 	migrationsv1 "kubevirt.io/api/migrations/v1alpha1"
 
@@ -72,7 +74,6 @@ import (
 	expect "github.com/google/goexpect"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/onsi/gomega/gstruct"
 	k8sv1 "k8s.io/api/core/v1"
 	policyv1 "k8s.io/api/policy/v1"
 	policyv1beta1 "k8s.io/api/policy/v1beta1"
@@ -120,7 +121,7 @@ const (
 	stressDefaultSleepDuration = 1600
 )
 
-var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system][sig-compute] VM Live Migration", decorators.SigComputeMigrations, decorators.SigCompute, func() {
+var _ = SIGMigrationDescribe("VM Live Migration", func() {
 	var virtClient kubecli.KubevirtClient
 	var migrationBandwidthLimit resource.Quantity
 	var err error
@@ -318,11 +319,13 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 		vmiSelector := v1.AppLabel + "=virt-launcher"
 		k8sClient := clientcmd.GetK8sCmdClient()
 		if k8sClient == "oc" {
-			clientcmd.RunCommandWithNS("", k8sClient, "adm", "drain", node, "--delete-emptydir-data", "--pod-selector", vmiSelector,
+			_, _, err := clientcmd.RunCommandWithNS("", k8sClient, "adm", "drain", node, "--delete-emptydir-data", "--pod-selector", vmiSelector,
 				"--ignore-daemonsets=true", "--force", "--timeout=180s")
+			Expect(err).ToNot(HaveOccurred())
 		} else {
-			clientcmd.RunCommandWithNS("", k8sClient, "drain", node, "--delete-emptydir-data", "--pod-selector", vmiSelector,
+			_, _, err := clientcmd.RunCommandWithNS("", k8sClient, "drain", node, "--delete-emptydir-data", "--pod-selector", vmiSelector,
 				"--ignore-daemonsets=true", "--force", "--timeout=180s")
+			Expect(err).ToNot(HaveOccurred())
 		}
 	}
 
@@ -365,196 +368,6 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 	BeforeEach(func() {
 		checks.SkipIfMigrationIsNotPossible()
 	})
-
-	confirmVMIPostMigrationFailed := func(vmi *v1.VirtualMachineInstance, migrationUID string) {
-		By("Retrieving the VMI post migration")
-		vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
-		Expect(err).ToNot(HaveOccurred())
-
-		By("Verifying the VMI's migration state")
-		Expect(vmi.Status.MigrationState).ToNot(BeNil())
-		Expect(vmi.Status.MigrationState.StartTimestamp).ToNot(BeNil())
-		Expect(vmi.Status.MigrationState.EndTimestamp).ToNot(BeNil())
-		Expect(vmi.Status.MigrationState.SourceNode).To(Equal(vmi.Status.NodeName))
-		Expect(vmi.Status.MigrationState.TargetNode).ToNot(Equal(vmi.Status.MigrationState.SourceNode))
-		Expect(vmi.Status.MigrationState.Completed).To(BeTrue())
-		Expect(vmi.Status.MigrationState.Failed).To(BeTrue())
-		Expect(vmi.Status.MigrationState.TargetNodeAddress).ToNot(Equal(""))
-		Expect(string(vmi.Status.MigrationState.MigrationUID)).To(Equal(migrationUID))
-
-		By("Verifying the VMI's is in the running state")
-		Expect(vmi.Status.Phase).To(Equal(v1.Running))
-	}
-	confirmVMIPostMigrationAborted := func(vmi *v1.VirtualMachineInstance, migrationUID string, timeout int) *v1.VirtualMachineInstance {
-		By("Waiting until the migration is completed")
-		EventuallyWithOffset(1, func() v1.VirtualMachineInstanceMigrationState {
-			vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
-			ExpectWithOffset(2, err).ToNot(HaveOccurred())
-
-			if vmi.Status.MigrationState != nil {
-				return *vmi.Status.MigrationState
-			}
-			return v1.VirtualMachineInstanceMigrationState{}
-
-		}, timeout, 1*time.Second).Should(
-			gstruct.MatchFields(gstruct.IgnoreExtras, gstruct.Fields{
-				"Completed":   BeTrue(),
-				"AbortStatus": Equal(v1.MigrationAbortSucceeded),
-			}),
-		)
-
-		By("Retrieving the VMI post migration")
-		vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
-		ExpectWithOffset(1, err).ToNot(HaveOccurred())
-
-		By("Verifying the VMI's migration state")
-		ExpectWithOffset(1, vmi.Status.MigrationState).ToNot(BeNil())
-		ExpectWithOffset(1, vmi.Status.MigrationState.StartTimestamp).ToNot(BeNil())
-		ExpectWithOffset(1, vmi.Status.MigrationState.EndTimestamp).ToNot(BeNil())
-		ExpectWithOffset(1, vmi.Status.MigrationState.SourceNode).To(Equal(vmi.Status.NodeName))
-		ExpectWithOffset(1, vmi.Status.MigrationState.TargetNode).ToNot(Equal(vmi.Status.MigrationState.SourceNode))
-		ExpectWithOffset(1, vmi.Status.MigrationState.TargetNodeAddress).ToNot(Equal(""))
-		ExpectWithOffset(1, string(vmi.Status.MigrationState.MigrationUID)).To(Equal(migrationUID))
-		ExpectWithOffset(1, vmi.Status.MigrationState.Failed).To(BeTrue())
-		ExpectWithOffset(1, vmi.Status.MigrationState.AbortRequested).To(BeTrue())
-
-		By("Verifying the VMI's is in the running state")
-		ExpectWithOffset(1, vmi).To(BeInPhase(v1.Running))
-		return vmi
-	}
-
-	cancelMigration := func(migration *v1.VirtualMachineInstanceMigration, vminame string, with_virtctl bool) {
-		if !with_virtctl {
-			By("Cancelling a Migration")
-			Expect(virtClient.VirtualMachineInstanceMigration(migration.Namespace).Delete(migration.Name, &metav1.DeleteOptions{})).To(Succeed(), "Migration should be deleted successfully")
-		} else {
-			By("Cancelling a Migration with virtctl")
-			command := clientcmd.NewRepeatableVirtctlCommand("migrate-cancel", "--namespace", migration.Namespace, vminame)
-			Expect(command()).To(Succeed(), "should successfully migrate-cancel a migration")
-		}
-	}
-
-	runAndCancelMigration := func(migration *v1.VirtualMachineInstanceMigration, vmi *v1.VirtualMachineInstance, with_virtctl bool, timeout int) *v1.VirtualMachineInstanceMigration {
-		By("Starting a Migration")
-		Eventually(func() error {
-			migration, err = virtClient.VirtualMachineInstanceMigration(migration.Namespace).Create(migration, &metav1.CreateOptions{})
-			return err
-		}, timeout, 1*time.Second).ShouldNot(HaveOccurred())
-
-		By("Waiting until the Migration is Running")
-
-		Eventually(func() bool {
-			migration, err := virtClient.VirtualMachineInstanceMigration(migration.Namespace).Get(migration.Name, &metav1.GetOptions{})
-			Expect(err).ToNot(HaveOccurred())
-
-			Expect(migration.Status.Phase).ToNot(Equal(v1.MigrationFailed))
-			if migration.Status.Phase == v1.MigrationRunning {
-				vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
-				Expect(err).ToNot(HaveOccurred())
-				if vmi.Status.MigrationState.Completed != true {
-					return true
-				}
-			}
-			return false
-
-		}, timeout, 1*time.Second).Should(BeTrue())
-
-		cancelMigration(migration, vmi.Name, with_virtctl)
-
-		return migration
-	}
-
-	runAndImmediatelyCancelMigration := func(migration *v1.VirtualMachineInstanceMigration, vmi *v1.VirtualMachineInstance, with_virtctl bool, timeout int) *v1.VirtualMachineInstanceMigration {
-		By("Starting a Migration")
-		Eventually(func() error {
-			migration, err = virtClient.VirtualMachineInstanceMigration(migration.Namespace).Create(migration, &metav1.CreateOptions{})
-			return err
-		}, timeout, 1*time.Second).ShouldNot(HaveOccurred())
-
-		By("Waiting until the Migration is Running")
-		Eventually(func() bool {
-			migration, err := virtClient.VirtualMachineInstanceMigration(migration.Namespace).Get(migration.Name, &metav1.GetOptions{})
-			Expect(err).ToNot(HaveOccurred())
-			return migration.Status.Phase == v1.MigrationRunning
-		}, timeout, 1*time.Second).Should(BeTrue())
-
-		cancelMigration(migration, vmi.Name, with_virtctl)
-
-		return migration
-	}
-
-	runMigrationAndExpectFailure := func(migration *v1.VirtualMachineInstanceMigration, timeout int) string {
-		By("Starting a Migration")
-		Eventually(func() error {
-			migration, err = virtClient.VirtualMachineInstanceMigration(migration.Namespace).Create(migration, &metav1.CreateOptions{})
-			return err
-		}, timeout, 1*time.Second).ShouldNot(HaveOccurred())
-		By("Waiting until the Migration Completes")
-
-		uid := ""
-		Eventually(func() v1.VirtualMachineInstanceMigrationPhase {
-			migration, err := virtClient.VirtualMachineInstanceMigration(migration.Namespace).Get(migration.Name, &metav1.GetOptions{})
-			Expect(err).ToNot(HaveOccurred())
-
-			phase := migration.Status.Phase
-			Expect(phase).NotTo(Equal(v1.MigrationSucceeded))
-
-			uid = string(migration.UID)
-			return phase
-
-		}, timeout, 1*time.Second).Should(Equal(v1.MigrationFailed))
-		return uid
-	}
-
-	runMigrationAndCollectMigrationMetrics := func(vmi *v1.VirtualMachineInstance, migration *v1.VirtualMachineInstanceMigration) {
-		var pod *k8sv1.Pod
-		var metricsIPs []string
-		const family = k8sv1.IPv4Protocol
-
-		By("Finding the prometheus endpoint")
-		pod, err = libnode.GetVirtHandlerPod(virtClient, vmi.Status.NodeName)
-		Expect(err).ToNot(HaveOccurred(), "Should find the virt-handler pod")
-		Expect(pod.Status.PodIPs).ToNot(BeEmpty(), "pod IPs must not be empty")
-		for _, ip := range pod.Status.PodIPs {
-			metricsIPs = append(metricsIPs, ip.IP)
-		}
-
-		By("Waiting until the Migration Completes")
-		ip := libinfra.GetSupportedIP(metricsIPs, family)
-
-		_ = tests.RunMigration(virtClient, migration)
-
-		By("Scraping the Prometheus endpoint")
-		validateNoZeroMetrics := func(metrics map[string]float64) error {
-			By("Checking the collected metrics")
-			keys := libinfra.GetKeysFromMetrics(metrics)
-			for _, key := range keys {
-				value := metrics[key]
-				if value == 0 {
-					return fmt.Errorf("metric value for %s is not expected to be zero", key)
-				}
-			}
-			return nil
-		}
-
-		getKubevirtVMMetricsFunc := tests.GetKubevirtVMMetricsFunc(&virtClient, pod)
-		Eventually(func() error {
-			out := getKubevirtVMMetricsFunc(ip)
-			lines := libinfra.TakeMetricsWithPrefix(out, "kubevirt_migrate_vmi")
-			metrics, err := libinfra.ParseMetricsToMap(lines)
-			Expect(err).ToNot(HaveOccurred())
-
-			if len(metrics) == 0 {
-				return fmt.Errorf("no metrics with kubevirt_migrate_vmi prefix are found")
-			}
-
-			if err := validateNoZeroMetrics(metrics); err != nil {
-				return err
-			}
-
-			return nil
-		}, 100*time.Second, 1*time.Second).ShouldNot(HaveOccurred())
-	}
 
 	runStressTest := func(vmi *v1.VirtualMachineInstance, vmsize string, stressTimeoutSeconds int) {
 		By("Run a stress test to dirty some pages and slow down the migration")
@@ -640,8 +453,8 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 			By("Executing a migration")
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-			migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
-			tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+			migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 			assertConnectivityToService("Asserting connectivity through service after migration")
 
@@ -676,8 +489,8 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			disks[len(disks)-1].Serial = secretDiskSerial
 
 			if migrationPolicy != nil {
-				tests.AlignPolicyAndVmi(vmi, migrationPolicy)
-				migrationPolicy = tests.CreateMigrationPolicy(virtClient, migrationPolicy)
+				AlignPolicyAndVmi(vmi, migrationPolicy)
+				migrationPolicy = CreateMigrationPolicy(virtClient, migrationPolicy)
 			}
 			vmi = tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 
@@ -695,9 +508,9 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			// execute a migration, wait for finalized state
 			By("Starting the Migration for iteration")
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-			migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 			By("Checking VMI, confirm migration state")
-			tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+			libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 			confirmMigrationMode(vmi, mode)
 
 			By("Is agent connected after migration")
@@ -770,10 +583,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 					By("starting the migration")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+					migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 					// check VMI, confirm migration state
-					tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+					libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 					vmi, err := virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
 					Expect(err).ToNot(HaveOccurred())
@@ -817,10 +630,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -843,10 +656,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				runningVMISpec, err := tests.GetRunningVMIDomainSpec(vmi)
 				Expect(err).ToNot(HaveOccurred())
@@ -880,10 +693,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 			})
 
 			It("should migrate vmi with LiveMigrateIfPossible eviction strategy", func() {
@@ -902,10 +715,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 			})
 
 			It("should migrate vmi and use Live Migration method with read-only disks", func() {
@@ -922,10 +735,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				By("Ensuring migration is using Live Migration method")
 				Eventually(func() v1.VirtualMachineInstanceMigrationMethod {
@@ -946,9 +759,9 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				By("checking if the metrics are still updated after the migration")
 				Eventually(func() error {
@@ -1000,9 +813,9 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				By("Checking the TSC frequency on the Domain XML on the new node")
 				domainSpec, err = tests.GetRunningVMIDomainSpec(vmi)
@@ -1033,10 +846,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				By("checking that we really migrated a VMI with only the root bus")
 				domSpec, err := tests.GetRunningVMIDomainSpec(vmi)
@@ -1074,10 +887,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 			})
 
 			It("[test_id:1783]should be successfully migrated multiple times with cloud-init disk", func() {
@@ -1097,11 +910,11 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 					// execute a migration, wait for finalized state
 					By(fmt.Sprintf("Starting the Migration for iteration %d", i))
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+					migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 					// check VMI, confirm migration state
-					tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
-					tests.ConfirmMigrationDataIsStored(virtClient, migration, vmi)
+					libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
+					libmigration.ConfirmMigrationDataIsStored(virtClient, migration, vmi)
 
 					By("Check if Migrated VMI has updated IP and IPs fields")
 					Eventually(func() error {
@@ -1167,10 +980,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By(fmt.Sprintf("Starting the Migration"))
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -1195,16 +1008,16 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By(fmt.Sprintf("Starting the Migration"))
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// ensure the libvirt domain is persistent
 				persistent, err := libvirtDomainIsPersistent(virtClient, vmi)
 				Expect(err).ToNot(HaveOccurred(), "Should list libvirt domains successfully")
 				Expect(persistent).To(BeTrue(), "The VMI was not found in the list of libvirt persistent domains")
-				tests.EnsureNoMigrationMetadataInPersistentXML(vmi)
+				libmigration.EnsureNoMigrationMetadataInPersistentXML(vmi)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -1226,7 +1039,8 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				Expect(console.LoginToAlpine(vmi)).To(Succeed())
 
 				By("Pausing the VirtualMachineInstance")
-				virtClient.VirtualMachineInstance(vmi.Namespace).Pause(context.Background(), vmi.Name, &v1.PauseOptions{})
+				err := virtClient.VirtualMachineInstance(vmi.Namespace).Pause(context.Background(), vmi.Name, &v1.PauseOptions{})
+				Expect(err).ToNot(HaveOccurred())
 				Eventually(matcher.ThisVMI(vmi), 30*time.Second, 2*time.Second).Should(matcher.HaveConditionTrue(v1.VirtualMachineInstancePaused))
 
 				By("verifying that the vmi is still paused before migration")
@@ -1236,10 +1050,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				By("verifying that the vmi is still paused after migration")
 				isPaused, err := tests.LibvirtDomainIsPaused(virtClient, vmi)
@@ -1419,10 +1233,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("Starting the Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -1456,10 +1270,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("Starting the Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 				Eventually(matcher.ThisVMI(vmi), 12*time.Minute, 2*time.Second).Should(matcher.HaveConditionTrue(v1.VirtualMachineInstanceAgentConnected))
 
 				By("Checking that the migrated VirtualMachineInstance has an updated time")
@@ -1544,10 +1358,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Starting a Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -1594,7 +1408,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				}
 				wg.Wait()
 
-				tests.ExpectMigrationSuccess(virtClient, migration1, tests.MigrationWaitTime)
+				libmigration.ExpectMigrationSuccess(virtClient, migration1, tests.MigrationWaitTime)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -1620,10 +1434,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Starting a Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -1642,10 +1456,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				// execute a migration, wait for finalized state
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, 180)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, 180)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -1674,10 +1488,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("Starting the Migration for iteration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -1741,7 +1555,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Starting the Migration for iteration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				_ = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				_ = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				By("Agent stays connected")
 				Consistently(matcher.ThisVMI(vmi), 5*time.Minute, 10*time.Second).Should(matcher.HaveConditionTrue(v1.VirtualMachineInstanceAgentConnected))
@@ -1819,10 +1633,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Starting new migration and waiting for it to succeed")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, 340)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, 340)
 
 				By("Verifying Second Migration Succeeeds")
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				By("Checking that the launcher is running as qemu")
 				Expect(tests.GetIdOfLauncher(vmi)).To(Equal("107"))
@@ -1906,10 +1720,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Starting new migration and waiting for it to succeed")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, 340)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, 340)
 
 				By("Verifying Second Migration Succeeeds")
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				By("Checking that the launcher is running as qemu")
 				Expect(tests.GetIdOfLauncher(vmi)).To(Equal("0"))
@@ -1969,10 +1783,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 					By("starting the migration")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+					migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 					// check VMI, confirm migration state
-					tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+					libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 					// delete VMI
 					By("Deleting the VMI")
@@ -2069,7 +1883,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 					vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceMemory] = resource.MustParse(fedoraVMSize)
 
 					By("Limiting the bandwidth of migrations in the test namespace")
-					tests.CreateMigrationPolicy(virtClient, tests.PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
+					CreateMigrationPolicy(virtClient, PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
 
 					By("Starting the VirtualMachineInstance")
 					vmi = tests.RunVMIAndExpectLaunch(vmi, 240)
@@ -2195,8 +2009,8 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			Context("should migrate using for postcopy", func() {
 
 				applyMigrationPolicy := func(vmi *v1.VirtualMachineInstance) {
-					tests.AlignPolicyAndVmi(vmi, migrationPolicy)
-					migrationPolicy = tests.CreateMigrationPolicy(virtClient, migrationPolicy)
+					AlignPolicyAndVmi(vmi, migrationPolicy)
+					migrationPolicy = CreateMigrationPolicy(virtClient, migrationPolicy)
 				}
 
 				applyKubevirtCR := func() {
@@ -2240,10 +2054,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 					// execute a migration, wait for finalized state
 					By("Starting the Migration")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, 150)
+					migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, 150)
 
 					// check VMI, confirm migration state
-					tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+					libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 					confirmMigrationMode(vmi, v1.MigrationPostCopy)
 				},
 					Entry("a migration policy", applyWithMigrationPolicy),
@@ -2305,10 +2119,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 					// execute a migration, wait for finalized state
 					By("Starting the Migration")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					migrationUID := runMigrationAndExpectFailure(migration, 180)
+					migrationUID := libmigration.RunMigrationAndExpectFailure(migration, 180)
 
 					// check VMI, confirm migration state
-					confirmVMIPostMigrationFailed(vmi, migrationUID)
+					libmigration.ConfirmVMIPostMigrationFailed(vmi, migrationUID)
 
 					// delete VMI
 					By("Deleting the VMI")
@@ -2359,10 +2173,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("Starting the Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migrationUID := runMigrationAndExpectFailure(migration, 180)
+				migrationUID := libmigration.RunMigrationAndExpectFailure(migration, 180)
 
 				// check VMI, confirm migration state
-				confirmVMIPostMigrationFailed(vmi, migrationUID)
+				libmigration.ConfirmVMIPostMigrationFailed(vmi, migrationUID)
 
 				By("Removing our migration killer pods")
 				for _, podName := range createdPods {
@@ -2399,10 +2213,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Starting new migration and waiting for it to succeed")
 				migration = tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, 340)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, 340)
 
 				By("Verifying Second Migration Succeeeds")
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				// delete VMI
 				By("Deleting the VMI")
@@ -2427,10 +2241,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 					By("Starting the Migration")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
 					migration.Name = fmt.Sprintf("%s-iter-%d", vmi.Name, i)
-					migrationUID := runMigrationAndExpectFailure(migration, 180)
+					migrationUID := libmigration.RunMigrationAndExpectFailure(migration, 180)
 
 					// check VMI, confirm migration state
-					confirmVMIPostMigrationFailed(vmi, migrationUID)
+					libmigration.ConfirmVMIPostMigrationFailed(vmi, migrationUID)
 
 					Eventually(func() error {
 						vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
@@ -2472,10 +2286,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				// execute a migration, wait for finalized state
 				By("Starting the Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migrationUID := runMigrationAndExpectFailure(migration, 180)
+				migrationUID := libmigration.RunMigrationAndExpectFailure(migration, 180)
 
 				// check VMI, confirm migration state
-				confirmVMIPostMigrationFailed(vmi, migrationUID)
+				libmigration.ConfirmVMIPostMigrationFailed(vmi, migrationUID)
 
 				Eventually(func() error {
 					vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
@@ -2725,7 +2539,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceMemory] = resource.MustParse(fedoraVMSize)
 
 				By("Limiting the bandwidth of migrations in the test namespace")
-				tests.CreateMigrationPolicy(virtClient, tests.PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
+				CreateMigrationPolicy(virtClient, PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
 
 				By("Starting the VirtualMachineInstance")
 				vmi = tests.RunVMIAndExpectLaunch(vmi, 240)
@@ -2733,10 +2547,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				By("Starting the Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
 
-				migration = runAndCancelMigration(migration, vmi, with_virtctl, 180)
+				migration = libmigration.RunAndCancelMigration(migration, vmi, with_virtctl, 180)
 
 				// check VMI, confirm migration state
-				confirmVMIPostMigrationAborted(vmi, string(migration.UID), 180)
+				libmigration.ConfirmVMIPostMigrationAborted(vmi, string(migration.UID), 180)
 
 				By("Waiting for the migration object to disappear")
 				libwait.WaitForMigrationToDisappearWithTimeout(migration, 240)
@@ -2755,7 +2569,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceMemory] = resource.MustParse(fedoraVMSize)
 
 				By("Limiting the bandwidth of migrations in the test namespace")
-				tests.CreateMigrationPolicy(virtClient, tests.PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
+				CreateMigrationPolicy(virtClient, PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
 
 				By("Starting the VirtualMachineInstance")
 				vmi = tests.RunVMIAndExpectLaunch(vmi, 240)
@@ -2765,10 +2579,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				By("Starting the Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
 
-				migration = runAndImmediatelyCancelMigration(migration, vmi, with_virtctl, 60)
+				migration = libmigration.RunAndImmediatelyCancelMigration(migration, vmi, with_virtctl, 60)
 
 				// check VMI, confirm migration state
-				confirmVMIPostMigrationAborted(vmi, string(migration.UID), 60)
+				libmigration.ConfirmVMIPostMigrationAborted(vmi, string(migration.UID), 60)
 
 				By("Waiting for the target virt-launcher pod to disappear")
 				labelSelector, err := labels.Parse(fmt.Sprintf("%s=virt-launcher,%s=%s", v1.AppLabel, v1.CreatedByLabel, string(vmi.GetUID())))
@@ -2813,7 +2627,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceMemory] = resource.MustParse(fedoraVMSize)
 
 				By("Limiting the bandwidth of migrations in the test namespace")
-				tests.CreateMigrationPolicy(virtClient, tests.PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
+				CreateMigrationPolicy(virtClient, PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
 
 				By("Starting the VirtualMachineInstance")
 				vmi = tests.RunVMIAndExpectLaunch(vmi, 240)
@@ -2838,7 +2652,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				}, timeout, 1*time.Second).Should(BeTrue())
 
 				By("Cancelling migration")
-				cancelMigration(migration, vmi.Name, with_virtctl)
+				libmigration.CancelMigration(migration, vmi.Name, with_virtctl)
 
 				By("Waiting for the migration object to disappear")
 				libwait.WaitForMigrationToDisappearWithTimeout(migration, 240)
@@ -2894,7 +2708,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 					By("Trying to migrate VM and expect for the migration to get stuck")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					migration = tests.RunMigration(virtClient, migration)
+					migration = libmigration.RunMigration(virtClient, migration)
 					expectMigrationSchedulingPhase := func() v1.VirtualMachineInstanceMigrationPhase {
 						migration, err = virtClient.VirtualMachineInstanceMigration(migration.Namespace).Get(migration.Name, &metav1.GetOptions{})
 						Expect(err).ShouldNot(HaveOccurred())
@@ -3003,7 +2817,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			}
 
 			It("[test_id:6981]should migrate only to nodes supporting right cpu model", func() {
-				sourceNode, targetNode, err := getValidSourceNodeAndTargetNodeForHostModelMigration(virtClient)
+				sourceNode, targetNode, err := libmigration.GetValidSourceNodeAndTargetNodeForHostModelMigration(virtClient)
 				if err != nil {
 					Skip(err.Error())
 				}
@@ -3016,7 +2830,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				By("Creating a VMI with default CPU mode to land in source node")
 				vmi.Spec.Domain.CPU = &v1.CPU{Model: v1.CPUModeHostModel}
 				By("Making sure the vmi start running on the source node and will be able to run only in source/target nodes")
-				nodeAffinityRule, err := affinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
+				nodeAffinityRule, err := libmigration.AffinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
 				Expect(err).ToNot(HaveOccurred())
 				vmi.Spec.Affinity = &k8sv1.Affinity{
 					NodeAffinity: nodeAffinityRule,
@@ -3035,7 +2849,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Starting the migration and expecting it to end successfully")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				_ = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				_ = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 				By("Ensuring that target pod has correct nodeSelector label")
 				vmiPod := tests.GetRunningPodByVirtualMachineInstance(vmi, vmi.Namespace)
@@ -3068,12 +2882,12 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 					node, err = virtClient.CoreV1().Nodes().Get(context.Background(), vmi.Status.NodeName, metav1.GetOptions{})
 					Expect(err).ToNot(HaveOccurred())
 
-					node = stopNodeLabeller(node.Name, virtClient)
+					node = libinfra.StopNodeLabeller(node.Name, virtClient)
 				})
 
 				AfterEach(func() {
 					By("Resuming node labeller")
-					node = resumeNodeLabeller(node.Name, virtClient)
+					node = libinfra.ResumeNodeLabeller(node.Name, virtClient)
 					_, doesFakeHostLabelExists := node.Labels[fakeHostModelLabel]
 					Expect(doesFakeHostLabelExists).To(BeFalse(), fmt.Sprintf("label %s is expected to disappear from node %s", fakeHostModelLabel, node.Name))
 				})
@@ -3098,7 +2912,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 					By("Starting the migration")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					_ = tests.RunMigration(virtClient, migration)
+					_ = libmigration.RunMigration(virtClient, migration)
 
 					events.ExpectEvent(vmi, k8sv1.EventTypeWarning, watch.NoSuitableNodesForHostModelMigration)
 				})
@@ -3124,7 +2938,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 					vmi = tests.RunVMIAndExpectLaunch(vmi, 240)
 
 					for indx, node := range nodes {
-						patchedNode := stopNodeLabeller(node.Name, virtClient)
+						patchedNode := libinfra.StopNodeLabeller(node.Name, virtClient)
 						Expect(patchedNode).ToNot(BeNil())
 						nodes[indx] = *patchedNode
 					}
@@ -3133,7 +2947,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				AfterEach(func() {
 					By("Restore node to its original state")
 					for _, node := range nodes {
-						updatedNode := resumeNodeLabeller(node.Name, virtClient)
+						updatedNode := libinfra.ResumeNodeLabeller(node.Name, virtClient)
 
 						supportedHostModelLabelExists := false
 						for labelKey := range updatedNode.Labels {
@@ -3161,7 +2975,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 					By("Starting the migration")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					_ = tests.RunMigration(virtClient, migration)
+					_ = libmigration.RunMigration(virtClient, migration)
 
 					events.ExpectEvent(vmi, k8sv1.EventTypeWarning, watch.NoSuitableNodesForHostModelMigration)
 				})
@@ -3199,7 +3013,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 					By("Starting the Migration")
 					migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-					_ = tests.RunMigrationAndExpectCompletion(virtClient, migration, 180)
+					_ = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, 180)
 				},
 					Entry("a VMI annotation", setMigrationParallelismWithAnnotation),
 				)
@@ -3229,7 +3043,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				var expectedPolicyName *string
 				if defineMigrationPolicy {
 					By("Creating a migration policy that overrides cluster policy")
-					policy := tests.GeneratePolicyAndAlignVMI(vmi)
+					policy := GeneratePolicyAndAlignVMI(vmi)
 					policy.Spec.AllowAutoConverge = pointer.BoolPtr(false)
 
 					_, err := virtClient.MigrationPolicy().Create(context.Background(), policy, metav1.CreateOptions{})
@@ -3243,10 +3057,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Starting the Migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, 180)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, 180)
 
 				// check VMI, confirm migration state
-				tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				By("Retrieving the VMI post migration")
 				vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
@@ -3289,9 +3103,9 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("starting the migration")
 				migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-				migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
-				vmi = tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+				vmi = libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 				domXml, err = tests.GetRunningVirtualMachineInstanceDomainXML(virtClient, vmi)
 				Expect(err).ToNot(HaveOccurred())
@@ -3323,10 +3137,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			// execute a migration, wait for finalized state
 			By("Starting the Migration")
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-			migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 			// check VMI, confirm migration state
-			tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+			libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 			// delete VMI
 			By("Deleting the VMI")
@@ -3842,7 +3656,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceMemory] = resource.MustParse(fedoraVMSize)
 
 			By("Limiting the bandwidth of migrations in the test namespace")
-			tests.CreateMigrationPolicy(virtClient, tests.PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
+			CreateMigrationPolicy(virtClient, PreparePolicyAndVMIWithBandwidthLimitation(vmi, migrationBandwidthLimit))
 
 			By("Starting the VirtualMachineInstance")
 			vmi = tests.RunVMIAndExpectLaunch(vmi, 240)
@@ -3856,7 +3670,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			// execute a migration, wait for finalized state
 			By("Starting the Migration")
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-			runMigrationAndCollectMigrationMetrics(vmi, migration)
+			libmigration.RunMigrationAndCollectMigrationMetrics(vmi, migration)
 		})
 	})
 
@@ -3975,10 +3789,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 			By("starting the migration")
 			migration := tests.NewRandomMigration(hugepagesVmi.Name, hugepagesVmi.Namespace)
-			migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 			// check VMI, confirm migration state
-			tests.ConfirmVMIPostMigration(virtClient, hugepagesVmi, migration)
+			libmigration.ConfirmVMIPostMigration(virtClient, hugepagesVmi, migration)
 
 			// delete VMI
 			By("Deleting the VMI")
@@ -4013,7 +3827,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 			By("Performing a migration")
 			migration := tests.NewRandomMigration(cpuVMI.Name, cpuVMI.Namespace)
-			tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 		})
 		Context("and NUMA passthrough", func() {
 			It("should not make migrations fail", func() {
@@ -4038,7 +3852,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 				By("Performing a migration")
 				migration := tests.NewRandomMigration(cpuVMI.Name, cpuVMI.Namespace)
-				tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+				libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 			})
 		})
 	})
@@ -4077,7 +3891,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 		By("Performing a migration")
 		migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-		tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+		libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 		By("Verifying that all imageIDs are in a reproducible form on the target")
 		pod = tests.GetRunningPodByVirtualMachineInstance(vmi, vmi.Namespace)
@@ -4101,16 +3915,16 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 		const fakeHostModel = v1.HostModelCPULabel + "fakeHostModel"
 
 		BeforeEach(func() {
-			sourceNode, targetNode, err = getValidSourceNodeAndTargetNodeForHostModelMigration(virtClient)
+			sourceNode, targetNode, err = libmigration.GetValidSourceNodeAndTargetNodeForHostModelMigration(virtClient)
 			if err != nil {
 				Skip(err.Error())
 			}
-			targetNode = stopNodeLabeller(targetNode.Name, virtClient)
+			targetNode = libinfra.StopNodeLabeller(targetNode.Name, virtClient)
 		})
 
 		AfterEach(func() {
 			By("Resuming node labeller")
-			targetNode = resumeNodeLabeller(targetNode.Name, virtClient)
+			targetNode = libinfra.ResumeNodeLabeller(targetNode.Name, virtClient)
 
 			By("Validating that fake labels are being removed")
 			for _, labelKey := range []string{fakeRequiredFeature, fakeHostModel} {
@@ -4129,7 +3943,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			By("Creating a VMI with default CPU mode to land in source node")
 			vmiToMigrate.Spec.Domain.CPU = &v1.CPU{Model: v1.CPUModeHostModel}
 			By("Making sure the vmi start running on the source node and will be able to run only in source/target nodes")
-			nodeAffinityRule, err := affinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
+			nodeAffinityRule, err := libmigration.AffinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
 			Expect(err).ToNot(HaveOccurred())
 			vmiToMigrate.Spec.Affinity = &k8sv1.Affinity{
 				NodeAffinity: nodeAffinityRule,
@@ -4142,7 +3956,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			// execute a migration, wait for finalized state
 			By("Starting the Migration to target node(with the amazing feature")
 			migration := tests.NewRandomMigration(vmiToMigrate.Name, vmiToMigrate.Namespace)
-			tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 			vmiToMigrate, err = virtClient.VirtualMachineInstance(vmiToMigrate.Namespace).Get(context.Background(), vmiToMigrate.GetName(), &metav1.GetOptions{})
 			Expect(err).ShouldNot(HaveOccurred())
@@ -4159,7 +3973,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			}
 
 			By("Starting the Migration to return to the source node")
-			tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 			Expect(console.LoginToFedora(vmiToMigrate)).To(Succeed())
 
 			vmiToMigrate, err = virtClient.VirtualMachineInstance(vmiToMigrate.Namespace).Get(context.Background(), vmiToMigrate.GetName(), &metav1.GetOptions{})
@@ -4190,7 +4004,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			By("Creating a VMI with default CPU mode to land in source node")
 			vmiToMigrate.Spec.Domain.CPU = &v1.CPU{Model: v1.CPUModeHostModel}
 			By("Making sure the vmi start running on the source node and will be able to run only in source/target nodes")
-			nodeAffinityRule, err := affinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
+			nodeAffinityRule, err := libmigration.AffinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
 			Expect(err).ToNot(HaveOccurred())
 			vmiToMigrate.Spec.Affinity = &k8sv1.Affinity{
 				NodeAffinity: nodeAffinityRule,
@@ -4203,7 +4017,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			// execute a migration, wait for finalized state
 			By("Starting the Migration to target node(with the amazing feature")
 			migration := tests.NewRandomMigration(vmiToMigrate.Name, vmiToMigrate.Namespace)
-			tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 			vmiToMigrate, err = virtClient.VirtualMachineInstance(vmiToMigrate.Namespace).Get(context.Background(), vmiToMigrate.GetName(), &metav1.GetOptions{})
 			Expect(err).ShouldNot(HaveOccurred())
@@ -4402,8 +4216,8 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			libnode.AddLabelToNode(nodes[1].Name, testLabel1, "true")
 			cpuSetSource := getVirtLauncherCPUSet(vmi)
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-			migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
-			tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+			migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 			By("ensuring the target cpuset is different from the source")
 			vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
@@ -4426,19 +4240,19 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			virtClient = kubevirt.Client()
 
 			By("Creating the Network Attachment Definition")
-			nad := tests.GenerateMigrationCNINetworkAttachmentDefinition()
+			nad := libmigration.GenerateMigrationCNINetworkAttachmentDefinition()
 			_, err = virtClient.NetworkClient().K8sCniCncfIoV1().NetworkAttachmentDefinitions(flags.KubeVirtInstallNamespace).Create(context.TODO(), nad, metav1.CreateOptions{})
 			Expect(err).NotTo(HaveOccurred(), "Failed to create the Network Attachment Definition")
 
 			By("Setting it as the migration network in the KubeVirt CR")
-			tests.SetDedicatedMigrationNetwork(nad.Name)
+			libmigration.SetDedicatedMigrationNetwork(nad.Name)
 		})
 		AfterEach(func() {
 			By("Clearing the migration network in the KubeVirt CR")
-			tests.ClearDedicatedMigrationNetwork()
+			libmigration.ClearDedicatedMigrationNetwork()
 
 			By("Deleting the Network Attachment Definition")
-			nad := tests.GenerateMigrationCNINetworkAttachmentDefinition()
+			nad := libmigration.GenerateMigrationCNINetworkAttachmentDefinition()
 			err = virtClient.NetworkClient().K8sCniCncfIoV1().NetworkAttachmentDefinitions(flags.KubeVirtInstallNamespace).Delete(context.TODO(), nad.Name, metav1.DeleteOptions{})
 			Expect(err).NotTo(HaveOccurred(), "Failed to delete the Network Attachment Definition")
 		})
@@ -4452,10 +4266,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 			By("Starting the migration")
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-			migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 			By("Checking if the migration happened, and over the right network")
-			vmi = tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+			vmi = libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 			Expect(vmi.Status.MigrationState.TargetNodeAddress).To(HavePrefix("172.21.42."), "The migration did not appear to go over the dedicated migration network")
 
 			// delete VMI
@@ -4474,8 +4288,8 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 		By("Starting a Migration")
 		migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-		migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, 180)
-		tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+		migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, 180)
+		libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 
 		By("Ensuring MigrationConfiguration is updated")
 		vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
@@ -4590,12 +4404,12 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			By("Waiting for the migration to fail")
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
 			setEvacuationAnnotation(migration)
-			_ = runMigrationAndExpectFailure(migration, tests.MigrationWaitTime)
+			_ = libmigration.RunMigrationAndExpectFailure(migration, tests.MigrationWaitTime)
 
 			By("Try again")
 			migration = tests.NewRandomMigration(vmi.Name, vmi.Namespace)
 			setEvacuationAnnotation(migration)
-			_ = runMigrationAndExpectFailure(migration, tests.MigrationWaitTime)
+			_ = libmigration.RunMigrationAndExpectFailure(migration, tests.MigrationWaitTime)
 
 			events.ExpectEvent(vmi, k8sv1.EventTypeWarning, watch.MigrationBackoffReason)
 		})
@@ -4607,7 +4421,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			By("Waiting for the migration to fail")
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
 			setEvacuationAnnotation(migration)
-			_ = runMigrationAndExpectFailure(migration, tests.MigrationWaitTime)
+			_ = libmigration.RunMigrationAndExpectFailure(migration, tests.MigrationWaitTime)
 
 			By("Patch VMI")
 			patchBytes := []byte(fmt.Sprintf(`[{"op": "remove", "path": "/metadata/annotations/%s"}]`, patch.EscapeJSONPointer(v1.FuncTestForceLauncherMigrationFailureAnnotation)))
@@ -4617,7 +4431,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			By("Try again with backoff")
 			migration = tests.NewRandomMigration(vmi.Name, vmi.Namespace)
 			setEvacuationAnnotation(migration)
-			_ = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			_ = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 			// Intentionally modifying history
 			events.DeleteEvents(vmi, k8sv1.EventTypeWarning, watch.MigrationBackoffReason)
@@ -4625,7 +4439,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 			By("There should be no backoff now")
 			migration = tests.NewRandomMigration(vmi.Name, vmi.Namespace)
 			setEvacuationAnnotation(migration)
-			_ = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			_ = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 			By("Checking that no backoff event occurred")
 			events.ExpectNoEvent(vmi, k8sv1.EventTypeWarning, watch.MigrationBackoffReason)
@@ -4661,7 +4475,7 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 
 			By("Trying to migrate the VirtualMachineInstance")
 			migration := tests.NewRandomMigration(vmi.Name, testsuite.GetTestNamespace(vmi))
-			migration = tests.RunMigration(virtClient, migration)
+			migration = libmigration.RunMigration(virtClient, migration)
 			Eventually(func() *v1.VirtualMachineInstanceMigration {
 				migration, err := virtClient.VirtualMachineInstanceMigration(migration.Namespace).Get(migration.Name, &metav1.GetOptions{})
 				if err != nil {
@@ -4732,82 +4546,6 @@ func temporaryTLSConfig() *tls.Config {
 	}
 }
 
-func stopNodeLabeller(nodeName string, virtClient kubecli.KubevirtClient) *k8sv1.Node {
-	var err error
-	var node *k8sv1.Node
-
-	Expect(CurrentSpecReport().IsSerial).To(BeTrue(), "stopping / resuming node-labeller is supported for serial tests only")
-
-	By(fmt.Sprintf("Patching node to %s include %s label", nodeName, v1.LabellerSkipNodeAnnotation))
-	key, value := v1.LabellerSkipNodeAnnotation, "true"
-	libnode.AddAnnotationToNode(nodeName, key, value)
-
-	By(fmt.Sprintf("Expecting node %s to include %s label", nodeName, v1.LabellerSkipNodeAnnotation))
-	Eventually(func() bool {
-		node, err = virtClient.CoreV1().Nodes().Get(context.Background(), nodeName, metav1.GetOptions{})
-		Expect(err).ToNot(HaveOccurred())
-
-		value, exists := node.Annotations[v1.LabellerSkipNodeAnnotation]
-		return exists && value == "true"
-	}, 30*time.Second, time.Second).Should(BeTrue(), fmt.Sprintf("node %s is expected to have annotation %s", nodeName, v1.LabellerSkipNodeAnnotation))
-
-	return node
-}
-
-func resumeNodeLabeller(nodeName string, virtClient kubecli.KubevirtClient) *k8sv1.Node {
-	var err error
-	var node *k8sv1.Node
-
-	node, err = virtClient.CoreV1().Nodes().Get(context.Background(), nodeName, metav1.GetOptions{})
-	Expect(err).ToNot(HaveOccurred())
-
-	if _, isNodeLabellerStopped := node.Annotations[v1.LabellerSkipNodeAnnotation]; !isNodeLabellerStopped {
-		// Nothing left to do
-		return node
-	}
-
-	By(fmt.Sprintf("Patching node to %s not include %s annotation", nodeName, v1.LabellerSkipNodeAnnotation))
-	libnode.RemoveAnnotationFromNode(nodeName, v1.LabellerSkipNodeAnnotation)
-
-	// In order to make sure node-labeller has updated the node, the host-model label (which node-labeller
-	// makes sure always resides on any node) will be removed. After node-labeller is enabled again, the
-	// host model label would be expected to show up again on the node.
-	By(fmt.Sprintf("Removing host model label %s from node %s (so we can later expect it to return)", v1.HostModelCPULabel, nodeName))
-	for _, label := range node.Labels {
-		if strings.HasPrefix(label, v1.HostModelCPULabel) {
-			libnode.RemoveLabelFromNode(nodeName, label)
-		}
-	}
-
-	libinfra.WakeNodeLabellerUp(virtClient)
-
-	By(fmt.Sprintf("Expecting node %s to not include %s annotation", nodeName, v1.LabellerSkipNodeAnnotation))
-	Eventually(func() error {
-		node, err = virtClient.CoreV1().Nodes().Get(context.Background(), nodeName, metav1.GetOptions{})
-		Expect(err).ShouldNot(HaveOccurred())
-
-		_, exists := node.Annotations[v1.LabellerSkipNodeAnnotation]
-		if exists {
-			return fmt.Errorf("node %s is expected to not have annotation %s", node.Name, v1.LabellerSkipNodeAnnotation)
-		}
-
-		foundHostModelLabel := false
-		for labelKey := range node.Labels {
-			if strings.HasPrefix(labelKey, v1.HostModelCPULabel) {
-				foundHostModelLabel = true
-				break
-			}
-		}
-		if !foundHostModelLabel {
-			return fmt.Errorf("node %s is expected to have a label with %s prefix. this means node-labeller is not enabled for the node", nodeName, v1.HostModelCPULabel)
-		}
-
-		return nil
-	}, 30*time.Second, time.Second).ShouldNot(HaveOccurred())
-
-	return node
-}
-
 func libvirtDomainIsPersistent(virtClient kubecli.KubevirtClient, vmi *v1.VirtualMachineInstance) (bool, error) {
 	vmiPod := tests.GetRunningPodByVirtualMachineInstance(vmi, vmi.Namespace)
 
@@ -4852,103 +4590,6 @@ func withEvictionStrategy(evictionStrategy v1.EvictionStrategy) libvmi.Option {
 	return func(vmi *v1.VirtualMachineInstance) {
 		vmi.Spec.EvictionStrategy = &evictionStrategy
 	}
-}
-
-func getValidSourceNodeAndTargetNodeForHostModelMigration(virtCli kubecli.KubevirtClient) (sourceNode *k8sv1.Node, targetNode *k8sv1.Node, err error) {
-	getNodeHostRequiredFeatures := func(node *k8sv1.Node) (features []string) {
-		for key, _ := range node.Labels {
-			if strings.HasPrefix(key, v1.HostModelRequiredFeaturesLabel) {
-				features = append(features, strings.TrimPrefix(key, v1.HostModelRequiredFeaturesLabel))
-			}
-		}
-		return features
-	}
-	areFeaturesSupportedOnNode := func(node *k8sv1.Node, features []string) bool {
-		isFeatureSupported := func(feature string) bool {
-			for key, _ := range node.Labels {
-				if strings.HasPrefix(key, v1.CPUFeatureLabel) && strings.Contains(key, feature) {
-					return true
-				}
-			}
-			return false
-		}
-		for _, feature := range features {
-			if !isFeatureSupported(feature) {
-				return false
-			}
-		}
-
-		return true
-	}
-
-	var sourceHostCpuModel string
-
-	nodes := libnode.GetAllSchedulableNodes(virtCli)
-	Expect(err).ToNot(HaveOccurred(), "Should list compute nodes")
-	for _, potentialSourceNode := range nodes.Items {
-		for _, potentialTargetNode := range nodes.Items {
-			if potentialSourceNode.Name == potentialTargetNode.Name {
-				continue
-			}
-
-			sourceHostCpuModel = tests.GetNodeHostModel(&potentialSourceNode)
-			if sourceHostCpuModel == "" {
-				continue
-			}
-			supportedInTarget := false
-			for key, _ := range potentialTargetNode.Labels {
-				if strings.HasPrefix(key, v1.SupportedHostModelMigrationCPU) && strings.Contains(key, sourceHostCpuModel) {
-					supportedInTarget = true
-					break
-				}
-			}
-
-			if supportedInTarget == false {
-				continue
-			}
-			sourceNodeHostModelRequiredFeatures := getNodeHostRequiredFeatures(&potentialSourceNode)
-			if areFeaturesSupportedOnNode(&potentialTargetNode, sourceNodeHostModelRequiredFeatures) == false {
-				continue
-			}
-			return &potentialSourceNode, &potentialTargetNode, nil
-		}
-	}
-	return nil, nil, fmt.Errorf("couldn't find valid nodes for host-model migration")
-}
-
-func affinityToMigrateFromSourceToTargetAndBack(sourceNode *k8sv1.Node, targetNode *k8sv1.Node) (nodefiinity *k8sv1.NodeAffinity, err error) {
-	if sourceNode == nil || targetNode == nil {
-		return nil, fmt.Errorf("couldn't find valid nodes for host-model migration")
-	}
-	return &k8sv1.NodeAffinity{
-		RequiredDuringSchedulingIgnoredDuringExecution: &k8sv1.NodeSelector{
-			NodeSelectorTerms: []k8sv1.NodeSelectorTerm{
-				{
-					MatchExpressions: []k8sv1.NodeSelectorRequirement{
-						{
-							Key:      "kubernetes.io/hostname",
-							Operator: k8sv1.NodeSelectorOpIn,
-							Values:   []string{sourceNode.Name, targetNode.Name},
-						},
-					},
-				},
-			},
-		},
-		PreferredDuringSchedulingIgnoredDuringExecution: []k8sv1.PreferredSchedulingTerm{
-			{
-				Preference: k8sv1.NodeSelectorTerm{
-					MatchExpressions: []k8sv1.NodeSelectorRequirement{
-						{
-							Key:      "kubernetes.io/hostname",
-							Operator: k8sv1.NodeSelectorOpIn,
-							Values:   []string{sourceNode.Name},
-						},
-					},
-				},
-				Weight: 1,
-			},
-		},
-	}, nil
 }
 
 func filterRunningMigrations(migrations []v1.VirtualMachineInstanceMigration) []v1.VirtualMachineInstanceMigration {
