+++ tests/monitoring/BUILD.bazel
@@ -24,6 +24,7 @@ go_library(
         "//tests/framework/checks:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
+        "//tests/libmigration:go_default_library",
         "//tests/libnode:go_default_library",
         "//tests/libvmi:go_default_library",
         "//tests/libwait:go_default_library",
