+++ tests/network/BUILD.bazel
@@ -54,6 +54,7 @@ go_library(
         "//tests/framework/checks:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
+        "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
         "//tests/libnet/cluster:go_default_library",
         "//tests/libnet/job:go_default_library",
