+++ tests/swap_test.go
@@ -26,6 +26,8 @@ import (
 	"strings"
 	"time"
 
+	"kubevirt.io/kubevirt/tests/libmigration"
+
 	"kubevirt.io/kubevirt/tests/decorators"
 
 	"kubevirt.io/kubevirt/tests/framework/checks"
@@ -132,7 +134,7 @@ var _ = Describe("[Serial][sig-compute]SwapTest", Serial, decorators.SigCompute,
 	}
 
 	getAffinityForTargetNode := func(targetNode *v1.Node) (nodeAffinity *v1.Affinity, err error) {
-		nodeAffinityRuleForVmiToFill, err := affinityToMigrateFromSourceToTargetAndBack(targetNode, targetNode)
+		nodeAffinityRuleForVmiToFill, err := libmigration.AffinityToMigrateFromSourceToTargetAndBack(targetNode, targetNode)
 		return &v1.Affinity{
 			NodeAffinity: nodeAffinityRuleForVmiToFill,
 		}, err
@@ -154,7 +156,7 @@ var _ = Describe("[Serial][sig-compute]SwapTest", Serial, decorators.SigCompute,
 
 	Context("Migration to/from memory overcommitted nodes", decorators.SigComputeMigrations, func() {
 		It("Postcopy Migration of vmi that is dirtying(stress-ng) more memory than the source node's memory", func() {
-			sourceNode, targetNode, err := getValidSourceNodeAndTargetNodeForHostModelMigration(virtClient)
+			sourceNode, targetNode, err := libmigration.GetValidSourceNodeAndTargetNodeForHostModelMigration(virtClient)
 			Expect(err).ToNot(HaveOccurred(), "should be able to get valid source and target nodes for migartion")
 			totalMemKib := getTotalMemSizeInKib(*sourceNode)
 			availableMemSizeKib := getAvailableMemSizeInKib(*sourceNode)
@@ -179,7 +181,7 @@ var _ = Describe("[Serial][sig-compute]SwapTest", Serial, decorators.SigCompute,
 			vmiMemSizeMi := resource.MustParse(fmt.Sprintf("%dMi", int((float64(memToUseInTheVmKib)+float64(gigbytesInkib*2))/bytesInKib)))
 
 			vmi := tests.NewRandomFedoraVMI()
-			nodeAffinityRule, err := affinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
+			nodeAffinityRule, err := libmigration.AffinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
 			Expect(err).ToNot(HaveOccurred())
 			vmi.Spec.Affinity = &v1.Affinity{
 				NodeAffinity: nodeAffinityRule,
@@ -204,10 +206,10 @@ var _ = Describe("[Serial][sig-compute]SwapTest", Serial, decorators.SigCompute,
 			// execute a migration, wait for finalized state
 			By("Starting the Migration")
 			migration := tests.NewRandomMigration(vmi.Name, vmi.Namespace)
-			migration = tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime*2)
+			migration = libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime*2)
 
 			// check VMI, confirm migration state
-			tests.ConfirmVMIPostMigration(virtClient, vmi, migration)
+			libmigration.ConfirmVMIPostMigration(virtClient, vmi, migration)
 			confirmMigrationMode(vmi, virtv1.MigrationPostCopy)
 			Expect(console.LoginToFedora(vmi)).To(Succeed())
 
@@ -224,7 +226,7 @@ var _ = Describe("[Serial][sig-compute]SwapTest", Serial, decorators.SigCompute,
 		})
 
 		It("Migration of vmi to memory overcommited node", func() {
-			sourceNode, targetNode, err := getValidSourceNodeAndTargetNodeForHostModelMigration(virtClient)
+			sourceNode, targetNode, err := libmigration.GetValidSourceNodeAndTargetNodeForHostModelMigration(virtClient)
 			Expect(err).ToNot(HaveOccurred(), "should be able to get valid source and target nodes for migartion")
 			vmMemoryRequestkib := 512000
 			availableMemSizeKib := getAvailableMemSizeInKib(*targetNode)
@@ -255,7 +257,7 @@ var _ = Describe("[Serial][sig-compute]SwapTest", Serial, decorators.SigCompute,
 			ExpectWithOffset(1, err).ToNot(HaveOccurred())
 
 			vmiToMigrate := tests.NewRandomFedoraVMI()
-			nodeAffinityRule, err := affinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
+			nodeAffinityRule, err := libmigration.AffinityToMigrateFromSourceToTargetAndBack(sourceNode, targetNode)
 			Expect(err).ToNot(HaveOccurred())
 			vmiToMigrate.Spec.Affinity = &v1.Affinity{
 				NodeAffinity: nodeAffinityRule,
@@ -270,7 +272,7 @@ var _ = Describe("[Serial][sig-compute]SwapTest", Serial, decorators.SigCompute,
 			// execute a migration, wait for finalized state
 			By("Starting the Migration")
 			migration := tests.NewRandomMigration(vmiToMigrate.Name, vmiToMigrate.Namespace)
-			tests.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
+			libmigration.RunMigrationAndExpectCompletion(virtClient, migration, tests.MigrationWaitTime)
 
 			By("The workloads in the node should consume more memory than the memory size eventually.")
 			swapSizeKib := getSwapSizeInKib(*targetNode)
